-- ACBrTEFComum -- [*] propriedade TACBrTEFResp.NumeroLoteTransacao, modificada de Integer para Int64
-- ACBrTEFDPayGoWeb --
[+] Adicionada Propriedade, para expor a classe de API
    property PGWebAPI: TACBrTEFPGWebAPI (readonly)
[*] Mudança no mapeamento de TACBrTEFDRespPayGoWeb.ConteudoToProperty:
    PWINFO_REQNUM -> NumeroLoteTransacao
    PWINFO_AUTHCODE -> CodigoAutorizacaoTransacao
[-] Correção no método "CNC"
[+] Método CNC modificado para aproveitar as informações da Transação em memória,
    para evitar a abertura de janelas, desnecessárias...
[-] Correções em "VerificarTransacoesPendentesClass"
    - Permitindo que Exceptions ocorridos, sejam exibidos
    - Correção de A.V. na chamada de "DoExibeMsg"
    (por: DSA)


git-svn-id: svn://svn.code.sf.net/p/acbr/code/trunk2@20242 6e92efe7-b92a-0410-a9ec-f9e4e41bb3a6

+++ Fontes/ACBrTEFD/ACBrTEFComum.pas
@@ -245,7 +245,7 @@ type
     fpNomeOperadoraCelular: String;
     fpNSU: String;
     fpNSUTransacaoCancelada: String;
-    fpNumeroLoteTransacao: Integer;
+    fpNumeroLoteTransacao: Int64;
     fpNumeroRecargaCelular: String;
     fpQtdLinhasComprovante: Integer;
     fpQtdParcelas: Integer;
@@ -328,7 +328,7 @@ type
     property TransacaoAprovada: Boolean read GetTransacaoAprovada;
     property TipoTransacao: Integer read fpTipoTransacao;
     property CodigoAutorizacaoTransacao: String read fpCodigoAutorizacaoTransacao;
-    property NumeroLoteTransacao: Integer read fpNumeroLoteTransacao;
+    property NumeroLoteTransacao: Int64 read fpNumeroLoteTransacao;
     property DataHoraTransacaoHost: TDateTime read fpDataHoraTransacaoHost;
     property DataHoraTransacaoLocal: TDateTime read fpDataHoraTransacaoLocal;
     property TipoParcelamento: Integer read fpTipoParcelamento;
